test: extract model fallback test families
