+++ src/research/model-fallback-pipeline.test.ts
@@ -31,7 +31,6 @@ import {
 import { reconstructPdf } from './pdf'
 import { pdfVisualMatchCandidateId } from './pdf-visuals'
 import { buildStructDocument } from '../struct/from-reconstruction'
-
 const modelIdentity = {
   providerId: 'recorded-stub',
   modelId: 'candidate-picker',
@@ -53,147 +52,6 @@ describe('PDF model fallback production adapter', () => {
       reconstructPdf(await fixtureFile('adjudication-required.pdf')),
       reconstructPdf(await fixtureFile('visual-adjudication-required.pdf')),
     ])
-  })
-
-  it('leaves the legacy result unchanged when no model configuration exists', () => {
-    expect(adjudicationRequired).not.toHaveProperty('modelConsultations')
-    expect(modelDecisionPointsForPdf(adjudicationRequired)).toHaveLength(3)
-    expect(
-      new Set(
-        modelDecisionPointsForPdf(adjudicationRequired).map(
-          ({ decisionClass }) => decisionClass,
-        ),
-      ),
-    ).toEqual(
-      new Set([
-        MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch,
-        MODEL_FALLBACK_DECISION_CLASSES.readingOrderTie,
-      ]),
-    )
-  })
-
-  it('orders fallback decisions by code unit independent of locale collation', () => {
-    const expected = modelDecisionPointsForPdf(adjudicationRequired).map(
-      ({ decisionClass, decisionId }) => `${decisionClass}\u0000${decisionId}`,
-    )
-    const localeCompare = vi
-      .spyOn(String.prototype, 'localeCompare')
-      .mockImplementation(function (this: string, other: string) {
-        return this < String(other) ? 1 : this > String(other) ? -1 : 0
-      })
-    try {
-      expect(
-        modelDecisionPointsForPdf(adjudicationRequired).map(
-          ({ decisionClass, decisionId }) =>
-            `${decisionClass}\u0000${decisionId}`,
-        ),
-      ).toEqual(expected)
-    } finally {
-      localeCompare.mockRestore()
-    }
-  })
-
-  it('wires an explicitly disabled gate through reconstructPdf without consulting or resolving', async () => {
-    const consult = vi.fn(() => ({ candidateId: 'not-called' }))
-    const result = await reconstructPdf(
-      await fixtureFile('adjudication-required.pdf'),
-      undefined,
-      {
-        modelFallback: {
-          ...PRODUCTION_PDF_MODEL_FALLBACK_OPTIONS,
-          model: { identity: modelIdentity, consult },
-        },
-      },
-    )
-
-    expect(consult).not.toHaveBeenCalled()
-    expect(structuredClone(PRODUCTION_PDF_MODEL_FALLBACK_OPTIONS)).toEqual({
-      enabled: false,
-      ownerOptIn: false,
-    })
-    expect(withoutReceipt(result)).toEqual(adjudicationRequired)
-    expect(validateModelConsultationReceipt(result.modelConsultations)).toBe(
-      true,
-    )
-    expect(result.modelConsultations).toMatchObject({
-      documentId: result.paper.id,
-      sourceSha256: result.source.sha256,
-      metrics: {
-        totalDecisionCount: 3,
-        totalConsultationCount: 0,
-        consultationRate: 0,
-      },
-    })
-    expect(result.diagnostics.map(({ code }) => code)).toEqual(
-      expect.arrayContaining([
-        'AMBIGUOUS_NOTE_MATCH',
-        'AMBIGUOUS_READING_ORDER',
-      ]),
-    )
-  })
-
-  it('consults only bounded candidates and replays note and reading-order choices without human provenance', async () => {
-    const requests: ModelDecisionRequest[] = []
-    const consult = vi.fn((request: ModelDecisionRequest) => {
-      requests.push(request)
-      return { candidateId: request.candidates[0]!.id }
-    })
-    const result = await reconstructPdf(
-      await fixtureFile('adjudication-required.pdf'),
-      undefined,
-      {
-        modelFallback: {
-          enabled: true,
-          ownerOptIn: true,
-          model: { identity: modelIdentity, consult },
-        },
-      },
-    )
-
-    expect(consult).toHaveBeenCalledTimes(3)
-    expect(result.modelConsultations?.metrics).toMatchObject({
-      totalDecisionCount: 3,
-      totalConsultationCount: 3,
-      consultationRate: 1,
-    })
-    expect(
-      result.noteRelationships.every(({ status }) => status === 'matched'),
-    ).toBe(true)
-    expect(result.diagnostics.map(({ code }) => code)).not.toEqual(
-      expect.arrayContaining([
-        'AMBIGUOUS_NOTE_MATCH',
-        'AMBIGUOUS_READING_ORDER',
-      ]),
-    )
-    expect(result.humanAdjudications).toEqual(
-      adjudicationRequired.humanAdjudications,
-    )
-    expect(JSON.stringify(requests)).not.toMatch(
-      /sourceText|altText|bytes|apiKey|accessToken|refreshToken/u,
-    )
-    expect(
-      result.noteRelationships.flatMap(({ evidence }) => evidence),
-    ).toContain('model-consultation')
-    expect(
-      result.noteRelationships.flatMap(({ evidence }) => evidence),
-    ).not.toContain('human-adjudication')
-    const readingRequest = requests.find(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.readingOrderTie,
-    )!
-    expect(readingRequest).toBeDefined()
-    expect(
-      new Set(
-        readingRequest.candidates.map((candidate) =>
-          JSON.stringify(candidate.regions),
-        ),
-      ).size,
-    ).toBe(readingRequest.candidates.length)
-    for (const candidate of readingRequest.candidates) {
-      expect(
-        (candidate.regions as Array<{ id: string }>).map(({ id }) => id),
-      ).toEqual(candidate.region_ids)
-    }
   })
 
   it('keeps abort ownership until asynchronous model fallback completes', async () => {
@@ -322,642 +180,6 @@ describe('PDF model fallback production adapter', () => {
     )
     expect(result.diagnostics.map(({ code }) => code)).not.toContain(
       'AMBIGUOUS_READING_ORDER',
-    )
-  })
-
-  it('resolves only the eligible visual ambiguity and keeps private source material out of requests and receipts', async () => {
-    const base = structuredClone(visualAdjudicationRequired)
-    const sentinel = 'PRIVATE_SOURCE_SENTINEL_8f9c2a'
-    for (const relationship of base.visualRelationships) {
-      relationship.sourceText = sentinel
-      relationship.altText = sentinel
-      for (const candidate of relationship.candidates) {
-        candidate.sourceText = sentinel
-        candidate.evidence.push(sentinel)
-      }
-    }
-    const requests: ModelDecisionRequest[] = []
-    const result = await resolvePdfModelFallbacks(base, {
-      enabled: true,
-      ownerOptIn: true,
-      distillation: new DistillationLedger(),
-      model: {
-        identity: modelIdentity,
-        consult: (request) => {
-          requests.push(request)
-          return { candidateId: request.candidates[0]!.id }
-        },
-      },
-    })
-
-    expect(requests).toHaveLength(1)
-    expect(result.diagnostics.map(({ code }) => code)).not.toContain(
-      'AMBIGUOUS_VISUAL_MATCH',
-    )
-    expect(result.diagnostics.map(({ code }) => code)).toContain(
-      'UNRESOLVED_VISUAL_OBJECT',
-    )
-    expect(
-      result.visualRelationships
-        .filter(({ status }) => status === 'matched')
-        .flatMap(({ evidence }) => evidence),
-    ).toContain('model-consultation')
-    const persistedEvidence = JSON.stringify({
-      requests,
-      receipt: result.modelConsultations,
-    })
-    expect(persistedEvidence).not.toContain(sentinel)
-    expect(persistedEvidence).not.toMatch(/sourceText|altText|bytes/u)
-    const [request] = requests
-    expect(request?.candidates).not.toHaveLength(0)
-    for (const candidate of request?.candidates ?? []) {
-      const evidenceCodes = candidate.evidence_codes as string[]
-      expect(evidenceCodes).toEqual([...new Set(evidenceCodes)].sort())
-    }
-  })
-
-  it('binds a visual consultation to the selected candidate kind', async () => {
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      {
-        enabled: true,
-        ownerOptIn: true,
-        distillation: new DistillationLedger(),
-        model: {
-          identity: modelIdentity,
-          consult: (request) => ({ candidateId: request.candidates[0]!.id }),
-        },
-      },
-    )
-    const consultation = resolved.modelConsultations!.consultations.find(
-      ({ decisionClass, status }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation &&
-        status === 'accepted',
-    )!
-    const relationship = resolved.visualRelationships.find(
-      ({ id }) => id === consultation.decisionId,
-    )!
-    relationship.kind = relationship.kind === 'figure' ? 'table' : 'figure'
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds a visual consultation to the installed candidate geometry', async () => {
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      {
-        enabled: true,
-        ownerOptIn: true,
-        distillation: new DistillationLedger(),
-        model: {
-          identity: modelIdentity,
-          consult: (request) => ({ candidateId: request.candidates[0]!.id }),
-        },
-      },
-    )
-    const consultation = resolved.modelConsultations!.consultations.find(
-      ({ decisionClass, status }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation &&
-        status === 'accepted',
-    )!
-    const relationship = resolved.visualRelationships.find(
-      ({ id }) => id === consultation.decisionId,
-    )!
-    const selected = relationship.candidates.find(
-      (candidate) =>
-        candidate.sourceRegionIds.join() ===
-        relationship.sourceRegionIds.join(),
-    )!
-    const replacementBoxes: NormalizedSourceBox[] = []
-    relationship.sourceBoxes = replacementBoxes
-    selected.sourceBoxes = replacementBoxes
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds accepted note receipts to their anchor, threshold, and evidence', async () => {
-    const resolved = await resolvePdfModelFallbacks(adjudicationRequired, {
-      enabled: true,
-      ownerOptIn: true,
-      model: {
-        identity: modelIdentity,
-        consult: (request) => ({ candidateId: request.candidates[0]!.id }),
-      },
-    })
-    const consultation = resolved.modelConsultations!.consultations.find(
-      ({ decisionClass, status }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch &&
-        status === 'accepted',
-    )!
-    const relationship = resolved.noteRelationships.find(
-      ({ id }) => id === consultation.decisionId,
-    )!
-    expect(relationship.canonicalAnchor?.kind).toBe('node')
-    if (relationship.canonicalAnchor?.kind !== 'node') {
-      throw new Error('Expected a canonical node anchor for the note fixture')
-    }
-    const alternateRelationship = resolved.noteRelationships.find(
-      ({ id, canonicalAnchor }) =>
-        id !== relationship.id && canonicalAnchor?.kind === 'node',
-    )
-    expect(alternateRelationship?.canonicalAnchor?.kind).toBe('node')
-    if (alternateRelationship?.canonicalAnchor?.kind !== 'node') {
-      throw new Error('Expected a second canonical note anchor for the fixture')
-    }
-    const alternateAnchor = structuredClone(
-      alternateRelationship.canonicalAnchor,
-    )
-
-    for (const mutate of [
-      (copy: typeof resolved) => {
-        const current = copy.noteRelationships.find(
-          ({ id }) => id === relationship.id,
-        )!
-        current.referenceStart = alternateRelationship.referenceStart
-        current.referenceEnd = alternateRelationship.referenceEnd
-        current.canonicalAnchor = structuredClone(alternateAnchor)
-      },
-      (copy: typeof resolved) => {
-        copy.noteRelationships.find(
-          ({ id }) => id === relationship.id,
-        )!.referenceRegionId = 'retargeted-note-anchor'
-      },
-      (copy: typeof resolved) => {
-        const current = copy.noteRelationships.find(
-          ({ id }) => id === relationship.id,
-        )!
-        current.threshold = Math.max(0, current.threshold - 0.1)
-      },
-      (copy: typeof resolved) => {
-        const current = copy.noteRelationships.find(
-          ({ id }) => id === relationship.id,
-        )!
-        current.evidence = ['model-consultation']
-      },
-    ]) {
-      const tampered = structuredClone(resolved)
-      mutate(tampered)
-      tampered.modelConsultations!.semanticStateSha256 =
-        pdfModelConsultationSemanticStateSha256(
-          tampered,
-          tampered.modelConsultations!,
-        )
-      expect(modelConsultationReceiptMatchesPdfReconstruction(tampered)).toBe(
-        false,
-      )
-    }
-  })
-
-  it('binds an accepted visual receipt to its caption region anchor', async () => {
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      {
-        enabled: true,
-        ownerOptIn: true,
-        distillation: new DistillationLedger(),
-        model: {
-          identity: modelIdentity,
-          consult: (request) => ({ candidateId: request.candidates[0]!.id }),
-        },
-      },
-    )
-    const consultation = resolved.modelConsultations!.consultations.find(
-      ({ decisionClass, status }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation &&
-        status === 'accepted',
-    )!
-    const relationship = resolved.visualRelationships.find(
-      ({ id }) => id === consultation.decisionId,
-    )!
-    relationship.captionRegionId = 'retargeted-caption-region'
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds a deterministic visual decision to its caption region anchor', async () => {
-    const [point] = modelDecisionPointsForPdf(
-      visualAdjudicationRequired,
-    ).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'caption-anchor-binding-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const relationship = resolved.visualRelationships.find(({ evidence }) =>
-      evidence.includes('deterministic-distillation'),
-    )!
-    relationship.captionRegionId = 'retargeted-caption-region'
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds a deterministic visual decision to its pre-resolution confidence', async () => {
-    const [point] = modelDecisionPointsForPdf(
-      visualAdjudicationRequired,
-    ).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation,
-    )
-    const originalConfidence = point!.inputs.confidence as number
-    const changed = structuredClone(visualAdjudicationRequired)
-    changed.visualRelationships.find(
-      ({ id }) => id === point!.decisionId,
-    )!.confidence = originalConfidence === 0 ? 0.01 : originalConfidence - 0.01
-    const changedPoint = modelDecisionPointsForPdf(changed).find(
-      ({ decisionId }) => decisionId === point!.decisionId,
-    )!
-
-    expect(changedPoint.inputs.confidence).not.toBe(originalConfidence)
-    expect(changedPoint.candidates.map(({ id }) => id)).not.toEqual(
-      point!.candidates.map(({ id }) => id),
-    )
-
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'caption-input-confidence-binding-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const relationship = resolved.visualRelationships.find(({ evidence }) =>
-      evidence.includes('deterministic-distillation'),
-    )!
-
-    expect(relationship.resolutionInputConfidence).toBe(originalConfidence)
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      true,
-    )
-
-    const tampered = structuredClone(resolved)
-    tampered.visualRelationships.find(
-      ({ id }) => id === relationship.id,
-    )!.resolutionInputConfidence =
-      originalConfidence === 0 ? 0.01 : originalConfidence - 0.01
-    tampered.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        tampered,
-        tampered.modelConsultations!,
-      )
-    expect(modelConsultationReceiptMatchesPdfReconstruction(tampered)).toBe(
-      false,
-    )
-  })
-
-  it('binds a deterministic visual decision to the complete installed evidence', async () => {
-    const [point] = modelDecisionPointsForPdf(
-      visualAdjudicationRequired,
-    ).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'caption-installed-evidence-binding-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const relationship = resolved.visualRelationships.find(({ evidence }) =>
-      evidence.includes('deterministic-distillation'),
-    )!
-    const selected = relationship.candidates.find(
-      ({ score }) => score === relationship.confidence,
-    )!
-    const diagnosticMarker = relationship.evidence.find((code) =>
-      code.startsWith('deterministically-distilled-'),
-    )!
-    expect(selected.evidence.length).toBeGreaterThan(0)
-    expect(diagnosticMarker).toBeDefined()
-
-    for (const mutateEvidence of [
-      (evidence: string[]) =>
-        evidence.filter((code) => code !== selected.evidence[0]),
-      (evidence: string[]) =>
-        evidence.filter((code) => code !== diagnosticMarker),
-      (evidence: string[]) => [...evidence, 'model-consultation'],
-    ]) {
-      const tampered = structuredClone(resolved)
-      const current = tampered.visualRelationships.find(
-        ({ id }) => id === relationship.id,
-      )!
-      current.evidence = mutateEvidence(current.evidence)
-      tampered.modelConsultations!.semanticStateSha256 =
-        pdfModelConsultationSemanticStateSha256(
-          tampered,
-          tampered.modelConsultations!,
-        )
-
-      expect(modelConsultationReceiptMatchesPdfReconstruction(tampered)).toBe(
-        false,
-      )
-    }
-  })
-
-  it('commits complete visual candidate evidence before deterministic replay', async () => {
-    const [point] = modelDecisionPointsForPdf(
-      visualAdjudicationRequired,
-    ).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'caption-complete-evidence-commitment-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const relationship = resolved.visualRelationships.find(({ evidence }) =>
-      evidence.includes('deterministic-distillation'),
-    )!
-    const selected = relationship.candidates.find(
-      ({ score }) => score === relationship.confidence,
-    )!
-    const uncommittedSignal = 'retired-rule-never-evaluated-visual-signal'
-    expect(point!.candidates[0]!.evidence_codes).not.toContain(
-      uncommittedSignal,
-    )
-    selected.evidence.push(uncommittedSignal)
-    relationship.evidence.push(uncommittedSignal)
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds a deterministic visual decision to its selected kind', async () => {
-    const [point] = modelDecisionPointsForPdf(
-      visualAdjudicationRequired,
-    ).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'caption-kind-binding-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const relationship = resolved.visualRelationships.find(({ evidence }) =>
-      evidence.includes('deterministic-distillation'),
-    )!
-    relationship.kind = relationship.kind === 'figure' ? 'table' : 'figure'
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds a deterministic visual decision to its candidate score', async () => {
-    const [point] = modelDecisionPointsForPdf(
-      visualAdjudicationRequired,
-    ).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'caption-score-binding-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const relationship = resolved.visualRelationships.find(({ evidence }) =>
-      evidence.includes('deterministic-distillation'),
-    )!
-    const selected = relationship.candidates.find(
-      (candidate) => candidate.score === relationship.confidence,
-    )!
-    selected.score = Math.max(0, selected.score - 0.1)
-    relationship.confidence = selected.score
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds a deterministic visual decision to its source identifiers', async () => {
-    const [point] = modelDecisionPointsForPdf(
-      visualAdjudicationRequired,
-    ).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'caption-source-id-binding-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const relationship = resolved.visualRelationships.find(({ evidence }) =>
-      evidence.includes('deterministic-distillation'),
-    )!
-    const selected = relationship.candidates.find(
-      ({ score }) => score === relationship.confidence,
-    )!
-    selected.sourceRegionIds = ['retargeted-source-region']
-    relationship.sourceRegionIds = [...selected.sourceRegionIds]
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds a deterministic visual decision to the installed source text', async () => {
-    const [point] = modelDecisionPointsForPdf(
-      visualAdjudicationRequired,
-    ).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'caption-source-text-binding-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const relationship = resolved.visualRelationships.find(({ evidence }) =>
-      evidence.includes('deterministic-distillation'),
-    )!
-    // `visualModelCandidateId` hashes each candidate's own text, so rewriting
-    // only the installed text leaves the deterministic choice id unchanged.
-    expect(
-      relationship.candidates.every(
-        ({ sourceText }) => sourceText !== undefined,
-      ),
-    ).toBe(true)
-    relationship.sourceText = `${relationship.sourceText} rewritten`
-    const canonicalNode = resolved.paper.nodes.find(
-      (node) => node.id === relationship.canonicalNodeId,
-    )
-    if (canonicalNode?.type === 'figure')
-      canonicalNode.sourceText = relationship.sourceText
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds a custom deterministic visual decision to the complete candidate set', async () => {
-    const [point] = modelDecisionPointsForPdf(
-      visualAdjudicationRequired,
-    ).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'caption-complete-candidate-set-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const relationship = resolved.visualRelationships.find(({ evidence }) =>
-      evidence.includes('deterministic-distillation'),
-    )!
-    const selected = relationship.candidates.find(
-      ({ score }) => score === relationship.confidence,
-    )!
-    const unselected = relationship.candidates.find(
-      (candidate) => candidate !== selected,
-    )!
-    unselected.score = Math.max(0, unselected.score - 0.1)
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('automatically retires a fully covered caption class for normal strict opt-in options', async () => {
-    const consult = vi.fn((request: ModelDecisionRequest) => ({
-      candidateId: request.candidates[0]!.id,
-    }))
-
-    const result = await resolvePdfModelFallbacks(visualAdjudicationRequired, {
-      enabled: true,
-      ownerOptIn: true,
-      model: { identity: modelIdentity, consult },
-    })
-
-    expect(consult).not.toHaveBeenCalled()
-    expect(result.modelConsultations?.consultations).toEqual([])
-    expect(result.modelConsultations?.decisions).toEqual([
-      expect.objectContaining({
-        decisionClass: MODEL_FALLBACK_DECISION_CLASSES.captionAssociation,
-        outcome: 'deterministic',
-        consulted: false,
-        deterministicRuleId: PDF_CAPTION_UNIQUE_BOUNDED_DISTANCE_RULE_ID,
-      }),
-    ])
-    expect(result.modelConsultations?.metrics).toMatchObject({
-      totalDecisionCount: 1,
-      totalConsultationCount: 0,
-      consultationRate: 0,
-    })
-    expect(result.diagnostics.map(({ code }) => code)).not.toContain(
-      'AMBIGUOUS_VISUAL_MATCH',
     )
   })
 
@@ -3042,671 +2264,6 @@ describe('PDF model fallback production adapter', () => {
       } else {
         tampered.readingOrder.edges[selectedEdgeIndex]!.status = 'candidate'
       }
-      tampered.modelConsultations!.semanticStateSha256 =
-        pdfModelConsultationSemanticStateSha256(
-          tampered,
-          tampered.modelConsultations!,
-        )
-
-      expect(modelConsultationReceiptMatchesPdfReconstruction(tampered)).toBe(
-        false,
-      )
-    }
-  })
-
-  it('binds a note consultation to its candidate score and installed confidence', async () => {
-    const resolved = await resolvePdfModelFallbacks(adjudicationRequired, {
-      enabled: true,
-      ownerOptIn: true,
-      model: {
-        identity: modelIdentity,
-        consult: (request) => ({ candidateId: request.candidates[0]!.id }),
-      },
-    })
-    const consultation = resolved.modelConsultations!.consultations.find(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch,
-    )!
-    const relationship = resolved.noteRelationships.find(
-      ({ id }) => id === consultation.decisionId,
-    )!
-    relationship.confidence = Math.max(0, relationship.confidence - 0.1)
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds a note consultation to the installed candidate source boxes', async () => {
-    const resolved = await resolvePdfModelFallbacks(adjudicationRequired, {
-      enabled: true,
-      ownerOptIn: true,
-      model: {
-        identity: modelIdentity,
-        consult: (request) => ({ candidateId: request.candidates[0]!.id }),
-      },
-    })
-    const consultation = resolved.modelConsultations!.consultations.find(
-      ({ decisionClass, status }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch &&
-        status === 'accepted',
-    )!
-    const relationship = resolved.noteRelationships.find(
-      ({ id }) => id === consultation.decisionId,
-    )!
-    const replacementBoxes: NormalizedSourceBox[] = relationship.sourceBoxes
-      .length
-      ? []
-      : [
-          {
-            page: 999,
-            x: 0,
-            y: 0,
-            width: 1,
-            height: 1,
-            rotation: 0,
-            method: 'pdf-text',
-          },
-        ]
-    relationship.sourceBoxes = replacementBoxes
-    const selected = relationship.candidates.find(
-      ({ targetNoteId }) => targetNoteId === relationship.targetNoteId,
-    )!
-    selected.sourceBoxes = structuredClone(replacementBoxes)
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds an accepted note receipt to the complete current candidate set', async () => {
-    const resolved = await resolvePdfModelFallbacks(adjudicationRequired, {
-      enabled: true,
-      ownerOptIn: true,
-      model: {
-        identity: modelIdentity,
-        consult: (request) => ({ candidateId: request.candidates[0]!.id }),
-      },
-    })
-    const consultation = resolved.modelConsultations!.consultations.find(
-      ({ decisionClass, status }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch &&
-        status === 'accepted',
-    )!
-    const relationship = resolved.noteRelationships.find(
-      ({ id }) => id === consultation.decisionId,
-    )!
-    const unselected = relationship.candidates.find(
-      ({ targetNoteId }) => targetNoteId !== relationship.targetNoteId,
-    )!
-    unselected.score = Math.max(0, unselected.score - 0.1)
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds an accepted visual receipt to every candidate source text', async () => {
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      {
-        enabled: true,
-        ownerOptIn: true,
-        distillation: new DistillationLedger(),
-        model: {
-          identity: modelIdentity,
-          consult: (request) => ({ candidateId: request.candidates[0]!.id }),
-        },
-      },
-    )
-    const consultation = resolved.modelConsultations!.consultations.find(
-      ({ decisionClass, status }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation &&
-        status === 'accepted',
-    )!
-    const relationship = resolved.visualRelationships.find(
-      ({ id }) => id === consultation.decisionId,
-    )!
-    const selected = relationship.candidates.find(
-      ({ score }) => score === relationship.confidence,
-    )!
-    const unselected = relationship.candidates.find(
-      (candidate) => candidate !== selected,
-    )!
-    unselected.sourceText = `${unselected.sourceText ?? ''} changed`
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds an accepted visual receipt to the installed source text', async () => {
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      {
-        enabled: true,
-        ownerOptIn: true,
-        distillation: new DistillationLedger(),
-        model: {
-          identity: modelIdentity,
-          consult: (request) => ({ candidateId: request.candidates[0]!.id }),
-        },
-      },
-    )
-    const consultation = resolved.modelConsultations!.consultations.find(
-      ({ decisionClass, status }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation &&
-        status === 'accepted',
-    )!
-    const relationship = resolved.visualRelationships.find(
-      ({ id }) => id === consultation.decisionId,
-    )!
-    // Every candidate carries its own `sourceText`, so rewriting the installed
-    // text leaves the candidate set — and therefore every candidate id —
-    // untouched.
-    expect(
-      relationship.candidates.every(
-        ({ sourceText }) => sourceText !== undefined,
-      ),
-    ).toBe(true)
-    relationship.sourceText = `${relationship.sourceText} rewritten`
-    const canonicalNode = resolved.paper.nodes.find(
-      (node) => node.id === relationship.canonicalNodeId,
-    )
-    if (canonicalNode?.type === 'figure')
-      canonicalNode.sourceText = relationship.sourceText
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds an accepted visual receipt to the complete installed evidence', async () => {
-    const resolved = await resolvePdfModelFallbacks(
-      visualAdjudicationRequired,
-      {
-        enabled: true,
-        ownerOptIn: true,
-        distillation: new DistillationLedger(),
-        model: {
-          identity: modelIdentity,
-          consult: (request) => ({ candidateId: request.candidates[0]!.id }),
-        },
-      },
-    )
-    const consultation = resolved.modelConsultations!.consultations.find(
-      ({ decisionClass, status }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.captionAssociation &&
-        status === 'accepted',
-    )!
-    const relationship = resolved.visualRelationships.find(
-      ({ id }) => id === consultation.decisionId,
-    )!
-    const selected = relationship.candidates.find(
-      ({ score }) => score === relationship.confidence,
-    )!
-    const diagnosticMarker = relationship.evidence.find((code) =>
-      code.startsWith('model-consulted-'),
-    )!
-    expect(selected.evidence.length).toBeGreaterThan(0)
-    expect(diagnosticMarker).toBeDefined()
-
-    for (const removedEvidence of [selected.evidence[0]!, diagnosticMarker]) {
-      const tampered = structuredClone(resolved)
-      const current = tampered.visualRelationships.find(
-        ({ id }) => id === relationship.id,
-      )!
-      current.evidence = current.evidence.filter(
-        (code) => code !== removedEvidence,
-      )
-      tampered.modelConsultations!.semanticStateSha256 =
-        pdfModelConsultationSemanticStateSha256(
-          tampered,
-          tampered.modelConsultations!,
-        )
-
-      expect(modelConsultationReceiptMatchesPdfReconstruction(tampered)).toBe(
-        false,
-      )
-    }
-  })
-
-  it('binds a deterministic note decision to its candidate score', async () => {
-    const [point] = modelDecisionPointsForPdf(adjudicationRequired).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'note-marker-match-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      adjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const decision = resolved.modelConsultations!.decisions.find(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch,
-    )!
-    const relationship = resolved.noteRelationships.find(
-      ({ id }) => id === decision.decisionId,
-    )!
-    const selected = relationship.candidates.find(
-      ({ targetNoteId }) => targetNoteId === relationship.targetNoteId,
-    )!
-    selected.score = Math.max(0, selected.score - 0.1)
-    relationship.confidence = selected.score
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds a deterministic note decision to its pre-resolution confidence', async () => {
-    const [point] = modelDecisionPointsForPdf(adjudicationRequired).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch,
-    )
-    const originalConfidence = point!.inputs.confidence as number
-    const changed = structuredClone(adjudicationRequired)
-    changed.noteRelationships.find(
-      ({ id }) => id === point!.decisionId,
-    )!.confidence = originalConfidence === 0 ? 0.01 : originalConfidence - 0.01
-    const changedPoint = modelDecisionPointsForPdf(changed).find(
-      ({ decisionId }) => decisionId === point!.decisionId,
-    )!
-
-    expect(changedPoint.inputs.confidence).not.toBe(originalConfidence)
-    expect(changedPoint.candidates.map(({ id }) => id)).not.toEqual(
-      point!.candidates.map(({ id }) => id),
-    )
-
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'note-input-confidence-binding-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      adjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const decision = resolved.modelConsultations!.decisions.find(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch,
-    )!
-    const relationship = resolved.noteRelationships.find(
-      ({ id }) => id === decision.decisionId,
-    )!
-
-    expect(relationship.resolutionInputConfidence).toBe(originalConfidence)
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      true,
-    )
-
-    const tampered = structuredClone(resolved)
-    tampered.noteRelationships.find(
-      ({ id }) => id === relationship.id,
-    )!.resolutionInputConfidence =
-      originalConfidence === 0 ? 0.01 : originalConfidence - 0.01
-    tampered.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        tampered,
-        tampered.modelConsultations!,
-      )
-    expect(modelConsultationReceiptMatchesPdfReconstruction(tampered)).toBe(
-      false,
-    )
-  })
-
-  it('binds a deterministic note decision to the installed candidate evidence', async () => {
-    const [point] = modelDecisionPointsForPdf(adjudicationRequired).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'note-marker-evidence-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      adjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const decision = resolved.modelConsultations!.decisions.find(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch,
-    )!
-    const relationship = resolved.noteRelationships.find(
-      ({ id }) => id === decision.decisionId,
-    )!
-    // `updateNoteRelationship` installs the selected candidate's evidence
-    // beside the origin marker, and STRUCT publishes it as reader-visible
-    // signals. Dropping one signal leaves candidates and choice untouched.
-    expect(relationship.evidence).toContain('deterministic-distillation')
-    relationship.evidence = relationship.evidence.filter(
-      (code) => code !== 'label-exact',
-    )
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('commits complete note candidate evidence before deterministic replay', async () => {
-    const [point] = modelDecisionPointsForPdf(adjudicationRequired).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'note-complete-evidence-commitment-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      adjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const decision = resolved.modelConsultations!.decisions.find(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch,
-    )!
-    const relationship = resolved.noteRelationships.find(
-      ({ id }) => id === decision.decisionId,
-    )!
-    const selected = relationship.candidates.find(
-      ({ targetNoteId }) => targetNoteId === relationship.targetNoteId,
-    )!
-    const uncommittedSignal = 'retired-rule-never-evaluated-note-signal'
-    expect(point!.candidates[0]!.evidence_codes).not.toContain(
-      uncommittedSignal,
-    )
-    selected.evidence.push(uncommittedSignal)
-    relationship.evidence.push(uncommittedSignal)
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds a deterministic note decision to the complete candidate set', async () => {
-    const [point] = modelDecisionPointsForPdf(adjudicationRequired).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'note-complete-candidate-set-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      adjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const decision = resolved.modelConsultations!.decisions.find(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch,
-    )!
-    const relationship = resolved.noteRelationships.find(
-      ({ id }) => id === decision.decisionId,
-    )!
-    const unselected = relationship.candidates.find(
-      ({ targetNoteId }) => targetNoteId !== relationship.targetNoteId,
-    )!
-    unselected.score = Math.max(0, unselected.score - 0.1)
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds a deterministic note decision to its surviving ambiguity inputs', async () => {
-    const [point] = modelDecisionPointsForPdf(adjudicationRequired).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'note-ambiguity-input-binding-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      adjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const decision = resolved.modelConsultations!.decisions.find(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.noteMarkerMatch,
-    )!
-    const relationship = resolved.noteRelationships.find(
-      ({ id }) => id === decision.decisionId,
-    )!
-
-    for (const mutate of [
-      (copy: typeof resolved) => {
-        copy.noteRelationships.find(
-          ({ id }) => id === relationship.id,
-        )!.referenceRegionId = 'retargeted-note-reference-region'
-      },
-      (copy: typeof resolved) => {
-        const current = copy.noteRelationships.find(
-          ({ id }) => id === relationship.id,
-        )!
-        current.threshold = Math.max(0, current.threshold - 0.1)
-      },
-    ]) {
-      const tampered = structuredClone(resolved)
-      mutate(tampered)
-      tampered.modelConsultations!.semanticStateSha256 =
-        pdfModelConsultationSemanticStateSha256(
-          tampered,
-          tampered.modelConsultations!,
-        )
-
-      expect(modelConsultationReceiptMatchesPdfReconstruction(tampered)).toBe(
-        false,
-      )
-    }
-  })
-
-  it('binds a deterministic reading decision to its region evidence', async () => {
-    const [point] = modelDecisionPointsForPdf(adjudicationRequired).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.readingOrderTie,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'reading-order-region-evidence-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      adjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const selectedIds = point!.candidates[0]!.region_ids as string[]
-    const region = resolved.regions.find(({ id }) => id === selectedIds[0])!
-    region.column = region.column === 'left' ? 'right' : 'left'
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds a deterministic reading decision to its surviving ambiguity evidence', async () => {
-    const [point] = modelDecisionPointsForPdf(adjudicationRequired).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.readingOrderTie,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'reading-order-ambiguity-evidence-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      adjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    // A retired rule may key off the tie's evidence codes, so dropping one
-    // leaves a versioned rule that need not still make this choice.
-    const tie = resolved.readingOrder.resolutions.find(
-      ({ resolutionOrigin }) =>
-        resolutionOrigin === 'deterministic-distillation',
-    )!
-    expect(tie.evidence.length).toBeGreaterThan(1)
-    tie.evidence = tie.evidence.slice(1)
-    resolved.modelConsultations!.semanticStateSha256 =
-      pdfModelConsultationSemanticStateSha256(
-        resolved,
-        resolved.modelConsultations!,
-      )
-
-    expect(modelConsultationReceiptMatchesPdfReconstruction(resolved)).toBe(
-      false,
-    )
-  })
-
-  it('binds a deterministic reading decision to its distillation origin', async () => {
-    const [point] = modelDecisionPointsForPdf(adjudicationRequired).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.readingOrderTie,
-    )
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => point!.candidates[0]!.id,
-      'reading-order-origin-binding-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      adjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-
-    for (const replacementOrigin of [
-      'model-consultation',
-      'human-adjudication',
-    ] as const) {
-      const tampered = structuredClone(resolved)
-      const tie = tampered.readingOrder.resolutions.find(
-        ({ resolutionOrigin }) =>
-          resolutionOrigin === 'deterministic-distillation',
-      )!
-      tie.resolutionOrigin = replacementOrigin
-      tampered.modelConsultations!.semanticStateSha256 =
-        pdfModelConsultationSemanticStateSha256(
-          tampered,
-          tampered.modelConsultations!,
-        )
-
-      expect(modelConsultationReceiptMatchesPdfReconstruction(tampered)).toBe(
-        false,
-      )
-    }
-  })
-
-  it('binds a deterministic reading decision to its installed accepted edges', async () => {
-    const [point] = modelDecisionPointsForPdf(adjudicationRequired).filter(
-      ({ decisionClass }) =>
-        decisionClass === MODEL_FALLBACK_DECISION_CLASSES.readingOrderTie,
-    )
-    const selected = point!.candidates[0]!
-    const distillation = new DistillationLedger()
-    distillation.registerFixture(point!)
-    distillation.retireClass(
-      point!.decisionClass,
-      () => selected.id,
-      'reading-order-edge-binding-v1',
-    )
-    const resolved = await resolvePdfModelFallbacks(
-      adjudicationRequired,
-      new ModelConsultationGate({ distillation }),
-    )
-    const selectedIds = new Set(selected.region_ids as string[])
-    const installedEdge = resolved.readingOrder.edges.find(
-      ({ from, to }) => selectedIds.has(from) && selectedIds.has(to),
-    )!
-    expect(installedEdge.status).toBe('accepted')
-
-    for (const mutate of [
-      (copy: typeof resolved) => {
-        copy.readingOrder.edges = copy.readingOrder.edges.filter(
-          ({ id }) => id !== installedEdge.id,
-        )
-      },
-      (copy: typeof resolved) => {
-        copy.readingOrder.edges.find(
-          ({ id }) => id === installedEdge.id,
-        )!.status = 'candidate'
-      },
-    ]) {
-      const tampered = structuredClone(resolved)
-      mutate(tampered)
       tampered.modelConsultations!.semanticStateSha256 =
         pdfModelConsultationSemanticStateSha256(
           tampered,
